Move methods to separate files
