--- internal/dev_server/api/delete_dev_project.go
@@ -1,1 +1,0 @@
-package api
