--- internal/dev_server/api/get_dev_projects.go
@@ -1,1 +1,0 @@
-package api
